+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stylec/vite
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @stylec/compiler@0.0.11
+  - @stylec/cli@0.0.11
+
 ## 0.0.10
 
 ### Patch Changes